🔨 refactor: location title 재구현
