+++ src/widgets/map/styles/MapTitle.css
@@ -1,0 +1,76 @@
+.map__title {
+  display: flex;
+  flex-direction: column;
+
+  gap: 0;
+  width: 100%;
+}
+
+.map__title__description {
+  display: flex;
+  flex-direction: row;
+
+  align-items: center;
+
+  gap: 0.78vmax;
+
+  color: #e7c889;
+}
+
+.map__title__description hr {
+  border: none;
+  border-top: 0.1vmax solid #e7c889;
+  width: 4.17vmax;
+}
+
+.map__title__description p {
+  font-weight: 700;
+  font-size: 0.63vmax;
+  letter-spacing: 0.15em;
+}
+
+.map__title h2 {
+  font-family: 'BookendBatang';
+  font-weight: 700;
+  font-size: 3.13vmax;
+
+  color: #38271d;
+}
+
+@media (max-width: 1024px) {
+  .map__title__description {
+    gap: 1.95vmin;
+  }
+
+  .map__title__description hr {
+    border-top: 0.26vmin solid #e7c889;
+    width: 9.11vmin;
+  }
+
+  .map__title__description p {
+    font-size: 1.3vmin;
+  }
+
+  .map__title h2 {
+    font-size: 5.21vmin;
+  }
+}
+
+@media (max-width: 767px) {
+  .map__title__description {
+    gap: 4vmin;
+  }
+
+  .map__title__description hr {
+    border-top: 0.53vmin solid #e7c889;
+    width: 16vmin;
+  }
+
+  .map__title__description p {
+    font-size: 2.13vmin;
+  }
+
+  .map__title h2 {
+    font-size: 8vmin;
+  }
+}
