--- src/widgets/map/styles/Title.css
@@ -1,70 +1,0 @@
-.map__title {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 0.78vmax;
-}
-
-.map__title_text {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.map__title_text p {
-  font-weight: 700;
-  font-size: 0.78vmax;
-  letter-spacing: 0.05em;
-  color: var(--brown-800);
-}
-
-.map__title_text h2 {
-  font-weight: 700;
-  font-size: 2.34vmax;
-  letter-spacing: 0.025em;
-  color: var(--typo-titleB);
-}
-
-.map__title hr {
-  width: 3.65vmax;
-  border: none;
-  border-top: 0.18vmax solid var(--brown-800);
-}
-
-@media (max-width: 1024px) {
-  .map__title {
-    gap: 1.95vmin;
-  }
-
-  .map__title_text p {
-    font-size: 1.95vmin;
-  }
-
-  .map__title_text h2 {
-    font-size: 5.86vmin;
-  }
-
-  .map__title hr {
-    width: 9.11vmin;
-    border-top: 0.46vmin solid var(--brown-800);
-  }
-}
-
-@media (max-width: 767px) {
-  .map__title {
-    gap: 2.67vmin;
-  }
-
-  .map__title_text p {
-    font-size: 3.2vmin;
-  }
-
-  .map__title_text h2 {
-    font-size: 8vmin;
-  }
-
-  .map__title hr {
-    width: 16vmin;
-    border-top: 0.93vmin solid var(--brown-800);
-  }
-}
