+++ src/widgets/map/ui/MapTitle.tsx
@@ -1,0 +1,13 @@
+import '../styles/MapTitle.css';
+
+export function MapTitle() {
+  return (
+    <hgroup className='map__title'>
+      <div className='map__title__description'>
+        <hr />
+        <p>LOCATION</p>
+      </div>
+      <h2>찾아오시는 길</h2>
+    </hgroup>
+  );
+}
