--- src/widgets/map/ui/Title.tsx
@@ -1,13 +1,0 @@
-import '../styles/Title.css';
-
-export function MapTitle() {
-  return (
-    <div className='map__title'>
-      <hgroup className='map__title_text'>
-        <p>LOCATION</p>
-        <h2>찾아오시는 길</h2>
-      </hgroup>
-      <hr />
-    </div>
-  );
-}
